pyproject.toml
[project]
name = "PiBot"
version = "1.0.0a"
requires-python = ">=3.11"
dependencies = ["discord.py", "python-dotenv", "oracledb"]
authors = [{ name = "Punit Turlapati", email = "punitsst@gmail.com" }]
description = "A custom bot for the Nu Deuteron chapter of Alpha Epsilon Pi"
readme = "README.md"
license = { file = "LICENCE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
]

[project.optional-dependencies]
dev = ["mypy", "pytest"]

[project.urls]
Repository = "https://github.com/TheBobTheBlob/PiBot"


[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"


[tool.pytest.ini_options]
testpaths = ["tests"]
